+++ CHANGELOG.md
@@ -13,6 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Security
 
+## [1.7.5] - 2023-05-28
+### Added
+- SuccessMessageMixin to TeamNewEditBaseView
+### Changed
+- RaceDetailView to RaceTahminView
+- Refactor NewTahminView
+### Fixed
+- next() to next
+
 ## [1.7.4] - 2023-05-27
 ### Fixed
 - Star imports